[SPARK-57328][SQL] Extract coalesce hint-building helpers into CoalesceHintUtils
Refactor the COALESCE / REPARTITION / REPARTITION_BY_RANGE / REBALANCE
hint-building helpers out of ResolveHints.ResolveCoalesceHints into a reusable
CoalesceHintUtils object so the pure, stateless logic can be shared without
depending on fixed-point rule state. No behavior change.
